requirements.txt
Flask==2.3.3
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.4
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
# Telegram Bot Dependencies
python-telegram-bot==20.7
aiofiles==23.2.1
pandas==2.1.4
openpyxl==3.1.2
